app/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
PyMySQL==1.1.0
bcrypt==4.1.3
